requirements.txt
fastapi
uvicorn
pydantic
newspaper3k
lxml-html-clean
pillow
python-dotenv
streamlit